uncopling reusable components from home root component
