+++ src/components/Header/Header.js
@@ -1,0 +1,12 @@
+import styled from 'styled-components'
+
+const Header = styled.div`
+  background: black;
+  padding: 1rem;
+  display: flex;
+  flex-flow: column wrap;
+  align-items: center;
+  justify-content: center;
+`;
+
+export default Header;
