+++ src/components/Search/Search.js
@@ -1,0 +1,9 @@
+import styled from 'styled-components';
+
+const Search = styled.input`
+  border: none;
+  border-radius: 5px;
+  padding: .5rem;
+`;
+
+export default Search;
