+++ src/components/SearchButton/SearchButton.js
@@ -1,0 +1,10 @@
+import styled from 'styled-components';
+
+const SearchButton = styled.button`
+  background-color: #d50000;
+  border: none;
+  border-radius: 5px;
+  padding: .5rem;
+`;
+
+export default SearchButton;
