+++ src/components/Title/Title.js
@@ -1,0 +1,8 @@
+import styled from 'styled-components';
+
+const Title = styled.h1`
+  color: white;
+  padding: 1rem 0;
+`;
+
+export default Title;
